requirements.txt
python-chess==1.11.2
pygame==2.6.1